feat: add WebSocket server to broadcast mining progress and mined blocks
This commit adds a WebSocket server to the blockchain application to
broadcast mining progress and mined blocks to connected clients. The
`WsServer` class implements the `IWsServer` interface and provides a
`broadcast` method to send messages to all connected clients. The
`Block` class now has a `mineBlock` method that broadcasts mining
progress and mined blocks to the WebSocket server. The `Blockchain`
class now takes an instance of `WsServer` as a constructor argument and
passes it to the `Block` constructor. The `index.ts` file now creates an
instance of `WsServer` and passes it to the `Blockchain` constructor.