feat(miner.ts): add support for mining only one block
This commit adds a new feature to the miner script that allows the user to
choose whether to mine only one block or to keep mining indefinitely. The
user is prompted with a confirmation message asking if they want to mine
only one block. If the user chooses to mine only one block, the script will
exit after the block is mined. If the user chooses to keep mining, the
script will continue to mine blocks indefinitely.